Move apply_label_smoothing to losses
