+++ src/pykeen/losses.py
@@ -140,8 +140,6 @@ from class_resolver import Resolver, normalize_string
 from torch.nn import functional
 from torch.nn.modules.loss import _Loss
 
-from pykeen.utils import apply_label_smoothing
-
 __all__ = [
     # Base Classes
     'Loss',
@@ -161,6 +159,35 @@ __all__ = [
     # Utils
     'loss_resolver',
 ]
+
+
+def apply_label_smoothing(
+    labels: torch.FloatTensor,
+    epsilon: float,
+    num_classes: int,
+) -> torch.FloatTensor:
+    """Apply label smoothing to a target tensor.
+
+    Redistributes epsilon probability mass from the true target uniformly to the remaining classes by replacing
+        * a hard one by (1 - epsilon)
+        * a hard zero by epsilon / (num_classes - 1)
+
+    :param labels:
+        The one-hot label tensor.
+    :param epsilon:
+        The smoothing parameter. Determines how much probability should be transferred from the true class to the
+        other classes.
+    :param num_classes:
+        The number of classes.
+
+    ..seealso:
+        http://www.deeplearningbook.org/contents/regularization.html, chapter 7.5.1
+    """
+    new_label_true = (1.0 - epsilon)
+    new_label_false = epsilon / (num_classes - 1)
+    labels = new_label_true * labels + new_label_false * (1.0 - labels)
+    return labels
+
 
 _REDUCTION_METHODS = dict(
     mean=torch.mean,
