+++ src/pykeen/utils.py
@@ -86,7 +86,6 @@ __all__ = [
     'complex_normalize',
     'lp_norm',
     'powersum_norm',
-    'apply_label_smoothing',
 ]
 
 logger = logging.getLogger(__name__)
@@ -1088,31 +1087,3 @@ if __name__ == '__main__':
     import doctest
 
     doctest.testmod()
-
-
-def apply_label_smoothing(
-    labels: torch.FloatTensor,
-    epsilon: float,
-    num_classes: int,
-) -> torch.FloatTensor:
-    """Apply label smoothing to a target tensor.
-
-    Redistributes epsilon probability mass from the true target uniformly to the remaining classes by replacing
-        * a hard one by (1 - epsilon)
-        * a hard zero by epsilon / (num_classes - 1)
-
-    :param labels:
-        The one-hot label tensor.
-    :param epsilon:
-        The smoothing parameter. Determines how much probability should be transferred from the true class to the
-        other classes.
-    :param num_classes:
-        The number of classes.
-
-    ..seealso:
-        http://www.deeplearningbook.org/contents/regularization.html, chapter 7.5.1
-    """
-    new_label_true = (1.0 - epsilon)
-    new_label_false = epsilon / (num_classes - 1)
-    labels = new_label_true * labels + new_label_false * (1.0 - labels)
-    return labels
